requirements.txt
aiohttp==3.6.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-environ==0.4.5
idna==2.10
multidict==4.7.6
psycopg2-binary==2.8.6
pytz==2020.1
requests==2.24.0
slackclient==2.9.2
sqlparse==0.4.1
urllib3==1.25.10
yarl==1.5.1
